+++ es/core/helper/set-global-state.js
@@ -1,5 +1,5 @@
 import util from '../../support/util';
-import { BROADCAST_TRIGGERED_BY_CC_API_SET_GLOBAL_STATE, MODULE_GLOBAL } from '../../support/constant';
+import { BROADCAST_TRIGGERED_BY_CC_API_SET_GLOBAL_STATE, STATE_FOR_ALL_CC_INSTANCES_OF_ONE_MODULE, MODULE_GLOBAL } from '../../support/constant';
 import pickOneRef from './pick-one-ref';
 /****
  * if you are sure this state is really belong to global state, call cc.setGlobalState,