+++ es/core/register-to-default.js
@@ -1,0 +1,11 @@
+import register from './register';
+import { MODULE_DEFAULT } from '../support/constant';
+export default function (ccClassKey, option) {
+  if (option === void 0) {
+    option = {};
+  }
+
+  if (!option.sharedStateKeys) option.sharedStateKeys = 'all';
+  option.module = MODULE_DEFAULT;
+  register(ccClassKey, option);
+}